Release-1.1

+++ database/migrations/2020_11_12_204623_create_routine_logs_table.php
@@ -17,7 +17,7 @@ class CreateRoutineLogsTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('routine_id');
-            $table->integer('weight');
+            $table->float('weight', 8, 2);
             $table->integer('sets');
             $table->integer('reps');
